transfered register in userprofile view
