+++ django/transcendence/register/urls.py
@@ -7,8 +7,6 @@ urlpatterns = [
         path('api/login/', views.login),
         path('api/logout/', views.logout),
         path('api/is_token_valid/', views.is_token_valid),
-        path('api/userinfo/', views.user_info),
-        path('api/updateprofile/', views.update_profile),
         ]
 
 urlpatterns = format_suffix_patterns(urlpatterns)
