+++ django/transcendence/transcendence/urls.py
@@ -19,5 +19,6 @@ from django.urls import path, include
 
 urlpatterns = [
     path('', include('register.urls')),
+    path('', include('userprofile.urls')),
     path('api/admin/', admin.site.urls),
 ]
